+++ src/main/java/org/dromara/easyai/i/PsoFunction.java
@@ -3,5 +3,5 @@ package org.dromara.easyai.i;
 public interface PsoFunction {//粒子群回调函数
 
     //根据参数返回函数值
-    double getResult(double[] parameter,int id) throws Exception;
+    float getResult(float[] parameter,int id) throws Exception;
 }
